+++ src/main/java/com/github/theword/queqiao/tool/constant/ProtocolConstants.java
@@ -34,6 +34,10 @@ public final class ProtocolConstants {
     }
 
     public static final class Status {
+        /**
+         * 请求处理成功
+         */
+        public static final int SUCCESS = 200;
         public static final int BAD_REQUEST = 400;
         public static final int NOT_FOUND = 404;
         public static final int INTERNAL_ERROR = 500;